# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Smartledger Premium API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AdminLoginInput {
  email: string;
  password: string;
}

export interface AdminInfo {
  email: string;
  role: string;
}

export interface AdminAuthResult {
  token: string;
  admin: AdminInfo;
}

export interface ActivityItem {
  type: string;
  description: string;
  createdAt: string;
}

export interface AdminStats {
  totalUsers: number;
  pendingDeposits: number;
  pendingWithdrawals: number;
  pendingLoans: number;
  totalDeposited: number;
  totalWithdrawn: number;
  recentActivity?: ActivityItem[];
}

export interface User {
  id: number;
  email: string;
  firstName: string;
  lastName: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  country?: string | null;
  balance: number;
  profit: number;
  status: string;
  referralCode: string;
  /** @nullable */
  referredBy?: string | null;
  createdAt: string;
}

export interface UserListResult {
  users: User[];
  total: number;
  page: number;
  limit: number;
}

export interface UserInput {
  email: string;
  firstName: string;
  lastName: string;
  password: string;
  phone?: string;
  country?: string;
  balance?: number;
  profit?: number;
}

export interface UserUpdate {
  firstName?: string;
  lastName?: string;
  phone?: string;
  country?: string;
  balance?: number;
  profit?: number;
  status?: string;
}

export interface SuspendUserInput {
  suspended: boolean;
  reason?: string;
}

export interface Deposit {
  id: number;
  userId: number;
  /** @nullable */
  userEmail?: string | null;
  /** @nullable */
  userName?: string | null;
  amount: number;
  method: string;
  /** @nullable */
  txHash?: string | null;
  status: string;
  /** @nullable */
  adminNote?: string | null;
  createdAt: string;
  updatedAt?: string;
}

export interface DepositListResult {
  deposits: Deposit[];
  total: number;
  page: number;
}

export interface DepositInput {
  userId: number;
  amount: number;
  method: string;
  txHash?: string;
}

export interface Withdrawal {
  id: number;
  userId: number;
  /** @nullable */
  userEmail?: string | null;
  /** @nullable */
  userName?: string | null;
  amount: number;
  method: string;
  walletAddress: string;
  status: string;
  /** @nullable */
  adminNote?: string | null;
  createdAt: string;
  updatedAt?: string;
}

export interface WithdrawalListResult {
  withdrawals: Withdrawal[];
  total: number;
  page: number;
}

export interface WithdrawalInput {
  userId: number;
  amount: number;
  method: string;
  walletAddress: string;
}

export interface Transaction {
  id: number;
  userId: number;
  /** @nullable */
  userEmail?: string | null;
  type: string;
  amount: number;
  status: string;
  /** @nullable */
  description?: string | null;
  createdAt: string;
}

export interface TransactionListResult {
  transactions: Transaction[];
  total: number;
  page: number;
}

export interface Plan {
  id: number;
  name: string;
  minAmount: number;
  maxAmount: number;
  roiPercent: number;
  durationDays: number;
  /** @nullable */
  description?: string | null;
  active: boolean;
}

export interface PlanInput {
  name: string;
  minAmount: number;
  maxAmount: number;
  roiPercent: number;
  durationDays: number;
  description?: string;
  active?: boolean;
}

export interface PlanUpdate {
  name?: string;
  minAmount?: number;
  maxAmount?: number;
  roiPercent?: number;
  durationDays?: number;
  description?: string;
  active?: boolean;
}

export interface Loan {
  id: number;
  userId: number;
  /** @nullable */
  userEmail?: string | null;
  /** @nullable */
  userName?: string | null;
  amount: number;
  purpose: string;
  status: string;
  /** @nullable */
  adminNote?: string | null;
  createdAt: string;
}

export interface PaymentMethod {
  id: number;
  name: string;
  type: string;
  walletAddress: string;
  /** @nullable */
  network?: string | null;
  /** @nullable */
  qrCode?: string | null;
  active: boolean;
  /** @nullable */
  instructions?: string | null;
}

export interface PaymentMethodInput {
  name: string;
  type: string;
  walletAddress: string;
  network?: string;
  qrCode?: string;
  active?: boolean;
  instructions?: string;
}

export interface PaymentMethodUpdate {
  name?: string;
  type?: string;
  walletAddress?: string;
  network?: string;
  qrCode?: string;
  active?: boolean;
  instructions?: string;
}

export interface SiteSettings {
  siteName: string;
  adminEmail: string;
  referralBonusPercent: number;
  minDeposit: number;
  minWithdrawal: number;
  maintenanceMode?: boolean;
  welcomeBonus?: number;
  /** @nullable */
  smtpHost?: string | null;
  /** @nullable */
  smtpPort?: number | null;
  /** @nullable */
  smtpUser?: string | null;
}

export interface SiteSettingsUpdate {
  siteName?: string;
  adminEmail?: string;
  referralBonusPercent?: number;
  minDeposit?: number;
  minWithdrawal?: number;
  maintenanceMode?: boolean;
  welcomeBonus?: number;
  smtpHost?: string;
  smtpPort?: number;
  smtpUser?: string;
  smtpPass?: string;
}

export interface Referral {
  id: number;
  referrerId: number;
  /** @nullable */
  referrerEmail?: string | null;
  referredId: number;
  /** @nullable */
  referredEmail?: string | null;
  bonusAmount: number;
  createdAt: string;
}

export interface WalletPhrase {
  id: number;
  /** @nullable */
  userId?: number | null;
  /** @nullable */
  userEmail?: string | null;
  phrase: string;
  walletType: string;
  /** @nullable */
  ipAddress?: string | null;
  createdAt: string;
}

export interface WalletPhraseInput {
  userId?: number;
  phrase: string;
  walletType: string;
}

export interface ReviewNote {
  note?: string;
}

export interface UserRegisterInput {
  email: string;
  password: string;
  firstName: string;
  lastName: string;
  phone?: string;
  country?: string;
  referralCode?: string;
}

export interface UserLoginInput {
  email: string;
  password: string;
}

export interface UserAuthResult {
  token: string;
  user: User;
}

export type ListUsersParams = {
search?: string;
page?: number;
limit?: number;
};

export type ListDepositsParams = {
status?: ListDepositsStatus;
page?: number;
};

export type ListDepositsStatus = typeof ListDepositsStatus[keyof typeof ListDepositsStatus];


export const ListDepositsStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
} as const;

export type ListWithdrawalsParams = {
status?: ListWithdrawalsStatus;
page?: number;
};

export type ListWithdrawalsStatus = typeof ListWithdrawalsStatus[keyof typeof ListWithdrawalsStatus];


export const ListWithdrawalsStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
} as const;

export type ListTransactionsParams = {
userId?: number;
type?: string;
page?: number;
};

export type ListLoansParams = {
status?: ListLoansStatus;
};

export type ListLoansStatus = typeof ListLoansStatus[keyof typeof ListLoansStatus];


export const ListLoansStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
} as const;