# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Smartledger Premium API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/login:
    post:
      operationId: adminLogin
      tags: [auth]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAuthResult"
        "401":
          description: Invalid credentials
...
  /admin/me:
    get:
      operationId: getAdminMe
      tags: [auth]
      summary: Get current admin session
      responses:
        "200":
          description: Admin info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminInfo"
        "401":
          description: Not authenticated

  # --- Dashboard Stats ---
...
  /admin/deposits:
    get:
      operationId: listDeposits
      tags: [admin-deposits]
      summary: List deposits
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [pending, approved, rejected]
        - in: query
          name: page
          schema:
            type: integer
      responses:
        "200":
          description: Deposits list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DepositListResult"
  /admin/deposits/{id}:
    get:
      operationId: getDeposit
      tags: [admin-deposits]
      summary: Get deposit by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deposit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deposit"
  /admin/deposits/{id}/approve:
    post:
      operationId: approveDeposit
      tags: [admin-deposits]
      summary: Approve deposit
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewNote"
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deposit"
  /admin/deposits/{id}/reject:
    post:
      operationId: rejectDeposit
      tags: [admin-deposits]
      summary: Reject deposit
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewNote"
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deposit"

  # --- Withdrawals ---
...
  /admin/plans:
    get:
      operationId: listPlans
      tags: [admin-plans]
      summary: List investment plans
      responses:
        "200":
          description: Plans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Plan"
    post:
      operationId: createPlan
      tags: [admin-plans]
      summary: Create investment plan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlanInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plan"
  /admin/plans/{id}:
    patch:
      operationId: updatePlan
      tags: [admin-plans]
      summary: Update investment plan
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlanUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plan"
    delete:
      operationId: deletePlan
      tags: [admin-plans]
      summary: Delete investment plan
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # --- Loans ---
  /admin/loans:
    get:
      operationId: listLoans
      tags: [admin-loans]
      summary: List loan applications
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [pending, approved, rejected]
      responses:
        "200":
          description: Loans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Loan"
  /admin/loans/{id}/approve:
    post:
      operationId: approveLoan
      tags: [admin-loans]
      summary: Approve loan
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewNote"
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Loan"
  /admin/loans/{id}/reject:
    post:
      operationId: rejectLoan
      tags: [admin-loans]
      summary: Reject loan
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewNote"
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Loan"

  # --- Payment Methods ---
  /admin/payment-methods:
    get:
      operationId: listPaymentMethods
      tags: [admin-payment-methods]
      summary: List payment methods
      responses:
        "200":
          description: Payment methods
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PaymentMethod"
    post:
      operationId: createPaymentMethod
      tags: [admin-payment-methods]
      summary: Create payment method
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentMethodInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentMethod"
  /admin/payment-methods/{id}:
    patch:
      operationId: updatePaymentMethod
      tags: [admin-payment-methods]
      summary: Update payment method
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentMethodUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentMethod"
    delete:
      operationId: deletePaymentMethod
      tags: [admin-payment-methods]
      summary: Delete payment method
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # --- Settings ---
...
  /user/deposit:
    post:
      operationId: submitDeposit
      tags: [user-actions]
      summary: Submit a deposit request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepositInput"
      responses:
        "201":
          description: Deposit submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deposit"
...
components:
  schemas:
...
    AdminLoginInput:
      type: object
      required: [email, password]
      properties:
        email: { type: string }
        password: { type: string }
    AdminAuthResult:
      type: object
      required: [token, admin]
      properties:
        token: { type: string }
        admin:
          $ref: "#/components/schemas/AdminInfo"
    AdminInfo:
      type: object
      required: [email, role]
      properties:
        email: { type: string }
        role: { type: string }
...
    Deposit:
      type: object
      required: [id, userId, amount, method, status, createdAt]
      properties:
        id: { type: integer }
        userId: { type: integer }
        userEmail: { type: ["string", "null"] }
        userName: { type: ["string", "null"] }
        amount: { type: number }
        method: { type: string }
        txHash: { type: ["string", "null"] }
        status: { type: string }
        adminNote: { type: ["string", "null"] }
        createdAt: { type: string }
        updatedAt: { type: string }
    DepositListResult:
      type: object
      required: [deposits, total, page]
      properties:
        deposits:
          type: array
          items:
            $ref: "#/components/schemas/Deposit"
        total: { type: integer }
        page: { type: integer }
    DepositInput:
      type: object
      required: [userId, amount, method]
      properties:
        userId: { type: integer }
        amount: { type: number }
        method: { type: string }
        txHash: { type: string }
...
    Plan:
      type: object
      required: [id, name, minAmount, maxAmount, roiPercent, durationDays, active]
      properties:
        id: { type: integer }
        name: { type: string }
        minAmount: { type: number }
        maxAmount: { type: number }
        roiPercent: { type: number }
        durationDays: { type: integer }
        description: { type: ["string", "null"] }
        active: { type: boolean }
    PlanInput:
      type: object
      required: [name, minAmount, maxAmount, roiPercent, durationDays]
      properties:
        name: { type: string }
        minAmount: { type: number }
        maxAmount: { type: number }
        roiPercent: { type: number }
        durationDays: { type: integer }
        description: { type: string }
        active: { type: boolean }
    PlanUpdate:
      type: object
      properties:
        name: { type: string }
        minAmount: { type: number }
        maxAmount: { type: number }
        roiPercent: { type: number }
        durationDays: { type: integer }
        description: { type: string }
        active: { type: boolean }
    Loan:
      type: object
      required: [id, userId, amount, purpose, status, createdAt]
      properties:
        id: { type: integer }
        userId: { type: integer }
        userEmail: { type: ["string", "null"] }
        userName: { type: ["string", "null"] }
        amount: { type: number }
        purpose: { type: string }
        status: { type: string }
        adminNote: { type: ["string", "null"] }
        createdAt: { type: string }
    PaymentMethod:
      type: object
      required: [id, name, type, walletAddress, active]
      properties:
        id: { type: integer }
        name: { type: string }
        type: { type: string }
        walletAddress: { type: string }
        network: { type: ["string", "null"] }
        qrCode: { type: ["string", "null"] }
        active: { type: boolean }
        instructions: { type: ["string", "null"] }
    PaymentMethodInput:
      type: object
      required: [name, type, walletAddress]
      properties:
        name: { type: string }
        type: { type: string }
        walletAddress: { type: string }
        network: { type: string }
        qrCode: { type: string }
        active: { type: boolean }
        instructions: { type: string }
    PaymentMethodUpdate:
      type: object
      properties:
        name: { type: string }
        type: { type: string }
        walletAddress: { type: string }
        network: { type: string }
        qrCode: { type: string }
        active: { type: boolean }
        instructions: { type: string }
...
    ReviewNote:
      type: object
      properties:
        note: { type: string }
...