# lib/api-zod/src/generated/types/depositListResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Smartledger Premium API
 * OpenAPI spec version: 0.1.0
 */
import type { Deposit } from './deposit';

export interface DepositListResult {
  deposits: Deposit[];
  total: number;
  page: number;
}