# lib/api-zod/src/generated/types/planUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Smartledger Premium API
 * OpenAPI spec version: 0.1.0
 */

export interface PlanUpdate {
  name?: string;
  minAmount?: number;
  maxAmount?: number;
  roiPercent?: number;
  durationDays?: number;
  description?: string;
  active?: boolean;
}